day-15

+++ src/app/shared/base-layout/base-layout.component.ts
@@ -12,12 +12,14 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { CookieService } from 'ngx-cookie-service';
+import { ConfirmationService, ConfirmEventType, MessageService } from 'primeng/api';
 
 // component
 @Component({
   selector: 'app-base-layout',
   templateUrl: './base-layout.component.html',
-  styleUrls: ['./base-layout.component.css']
+  styleUrls: ['./base-layout.component.css'],
+  providers: [MessageService, ConfirmationService]
 })
 
 // export class
@@ -27,7 +29,8 @@ export class BaseLayoutComponent implements OnInit {
   year: number
 
   //  constructor
-  constructor(private cookieService: CookieService, private router: Router) {
+  constructor(private cookieService: CookieService, private router: Router,
+    private confirmationService: ConfirmationService, private messageService: MessageService) {
     this.sessionName = this.cookieService.get('session_name')
     this.year = Date.now()
   }
@@ -37,7 +40,24 @@ export class BaseLayoutComponent implements OnInit {
 
   // logout function
   logout() {
-    this.cookieService.deleteAll()
-    this.router.navigate(['/session/login'])
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to proceed?',
+      header: 'Log out confirmation',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.cookieService.deleteAll()
+        this.router.navigate(['/session/login'])
+      },
+      reject: (type: any) => {
+        switch (type) {
+          case ConfirmEventType.REJECT:
+            this.messageService.add({ severity: 'info', summary: 'Cancelled', detail: 'Log out cancelled' });
+            break
+            case ConfirmEventType.CANCEL:
+            this.messageService.add({ severity: 'info', summary: 'Cancelled', detail: 'Log out cancelled' });
+            break
+        }
+      }
+    })
   }
 }
